+++ src/tests/integration/already-authenticated.test.ts
@@ -1,12 +1,12 @@
-// import { createMocks } from './mocks'
-// import { runBackground } from './steps/run-background'
-// import { mountPopup } from './steps/mount-popup'
-// import { onceAuthenticated } from './steps/once-authenticated'
+import { createMocks } from './mocks'
+import { runBackground } from './steps/run-background'
+import { mountPopup } from './steps/mount-popup'
+import { onceAuthenticated } from './steps/once-authenticated'
 
-// describe('already authenticated when extension loads', () => {
-//   const mocks = createMocks()
+describe('already authenticated when extension loads', () => {
+  const mocks = createMocks()
 
-//   runBackground(mocks, { alreadyAuthenticated: true })
-//   mountPopup(mocks, { alreadyAuthenticated: true })
-//   onceAuthenticated(mocks)
-// })
+  runBackground(mocks, { alreadyAuthenticated: true })
+  mountPopup(mocks, { alreadyAuthenticated: true })
+  onceAuthenticated(mocks)
+})
